pyproject.toml
[project]
name = "projeto1"
version = "0.1.0"
description = "Ingestão e Busca Semântica com LangChain e Postgres"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"aiohappyeyeballs==2.6.1",
	"aiohttp==3.12.15",
	"aiosignal==1.4.0",
	"annotated-types==0.7.0",
	"anyio==4.10.0",
	"asyncpg==0.30.0",
	"attrs==25.3.0",
	"cachetools==5.5.2",
	"certifi==2025.8.3",
	"charset-normalizer==3.4.3",
	"dataclasses-json==0.6.7",
	"distro==1.9.0",
	"filetype==1.2.0",
	"frozenlist==1.7.0",
	"google-ai-generativelanguage==0.6.18",
	"google-api-core==2.25.1",
	"google-auth==2.40.3",
	"googleapis-common-protos==1.70.0",
	"greenlet==3.2.4",
	"grpcio==1.74.0",
	"grpcio-status==1.74.0",
	"h11==0.16.0",
	"httpcore==1.0.9",
	"httpx==0.28.1",
	"httpx-sse==0.4.1",
	"idna==3.10",
	"jiter==0.10.0",
	"jsonpatch==1.33",
	"jsonpointer==3.0.0",
	"langchain==0.3.27",
	"langchain-community==0.3.27",
	"langchain-core==0.3.74",
	"langchain-google-genai==2.1.9",
	"langchain-openai==0.3.30",
	"langchain-postgres==0.0.15",
	"langchain-text-splitters==0.3.9",
	"langsmith==0.4.20",
	"marshmallow==3.26.1",
	"multidict==6.6.4",
	"mypy_extensions==1.1.0",
	"numpy==2.3.2",
	"openai==1.102.0",
	"orjson==3.11.3",
	"packaging==25.0",
	"pgvector==0.3.6",
	"propcache==0.3.2",
	"proto-plus==1.26.1",
	"protobuf==6.32.0",
	"psycopg==3.2.9",
	"psycopg-binary==3.2.9",
	"psycopg-pool==3.2.6",
	"psycopg2-binary==2.9.10",
	"pyasn1==0.6.1",
	"pyasn1_modules==0.4.2",
	"pydantic==2.11.7",
	"pydantic-settings==2.10.1",
	"pydantic_core==2.33.2",
	"pypdf==6.0.0",
	"python-dotenv==1.1.1",
	"PyYAML==6.0.2",
	"regex==2025.7.34",
	"requests==2.32.5",
	"requests-toolbelt==1.0.0",
	"rsa==4.9.1",
	"sniffio==1.3.1",
	"SQLAlchemy==2.0.43",
	"tenacity==9.1.2",
	"tiktoken==0.11.0",
	"tqdm==4.67.1",
	"typing-inspect==0.9.0",
	"typing-inspection==0.4.1",
	"typing_extensions==4.15.0",
	"urllib3==2.5.0",
	"yarl==1.20.1",
	"zstandard==0.24.0"
]
